oversight fix and development text
i had previously put javascript:void(o) as the href for the "articles" dropdown box (which is supposed to do nothing) without realizing the o is supposed to be a 0.
There is also a line of text on the main page explaining the website's development and that it is unfinished.