index CSS responsiveness overhall + redesign
MAJOR: CSS should now be fully responsive to all viewport sizes (huge)

Improved the design of the social media link buttons and added a youtube button. additionally, added a new page that you are brought to by the discord button with my updated discord tag and a sarcastic blurb to express my frustrations.

changed some other styles of the site including removing the visibility of the <div> tile that houses all of the text and changing the font of the main header to a more artistic style.

I've realized now that, if nothing else, I'd like this website to serve as an artistic project created for fun.