jQuery + restructure + redesign
Added:
- jQuery package for future aid in scripting.
- Desktop-only "Other Links" dropdown for links unrelated to my social media and eventually for site-exclusive content
Revised/Removed:
- Various styling choices. Perhaps most notably, the site-wide background color is now rich black, and fonts are now more uniform.
- Flavor text. Particularly the Guy Debord quote from the index page and the discord page text.
- Twitch page now only contains 2 links to my main account and secondary account, and has reduced additional text. Subject to change in the future; I don't really know how I want that page to work.

 jQuery is NOT being loaded from a CDN because the benefits of such are not really that spectacular, and I couldn't find a reliable way to get VS Code to recognize jQuery type declarations in an external JS file without a node package.

I wish to continue the development of this website, hopefully with site-exclusive content in the future. Any creative work I produce will be available on the site in the future. I have vague plans in my head for games, applications, music, tools, essays, videos, books, streams, blog posts, matrix chat clients, a discord server/bot, and MORE to be available on the site.