Dependency localization/jQuery bugs fixed
Removed Node from the repository after I noticed the jQuery script was not loading, nor running properly when pushed to the live site. This is because running scripts with Node packages from website servers requires an extra step, of which I knew wasn't necessary as jQuery could be loaded more efficiently for my purposes, which is what this commit is adding.

jQuery SHOULD now be properly locally sourced from `\src\jquery.min.js` Will update soon after if not.

Decorative main header font is now also locally hosted, rather than being linked from the google fonts CDN previously. The license is also included, just in case. This was done for no other reason than a slight development annoyance: Whenever I would load the HTML file locally, for a very brief frame, the font of the main header would be that of default CSS, as it took some time to load the Google Font's file. This effect is not present on the live site, but I felt it was in theme with the jQuery changes, so I felt it wouldn't hurt.

Hosting the jQuery source file on the site's end because I don't like the idea of it being hosted on a 3rd party CDN elsewhere. I wrote in a previous commit (where I added Node modules) why I dislike depending on a CDN for jQuery.